+++ src/commands/start.ts
@@ -15,8 +15,7 @@
  */
 
 import { Command } from '@oclif/command';
-import { BootstrapService, BootstrapUtils } from '../service';
-import { CommandUtils } from '../service/CommandUtils';
+import { BootstrapService, BootstrapUtils, CommandUtils } from '../service';
 import Clean from './clean';
 import Compose from './compose';
 import Config from './config';
@@ -26,7 +25,7 @@ export default class Start extends Command {
     static description = 'Single command that aggregates config, compose and run in one line!';
 
     static examples = [
-        `$ symbol-bootstrap start`,
+        `$ symbol-bootstrap start -p bootstrap`,
         `$ symbol-bootstrap start -p bootstrap`,
         `$ symbol-bootstrap start -p testnet -a dual`,
         `$ symbol-bootstrap start -p testnet -a dual --password 1234`,