+++ src/commands/updateVotingKeys.ts
@@ -15,11 +15,9 @@
  */
 
 import { Command, flags } from '@oclif/command';
-import { join } from 'path';
 import { LogType } from '../logger';
 import Logger from '../logger/Logger';
 import LoggerFactory from '../logger/LoggerFactory';
-import { ConfigPreset } from '../model';
 import { BootstrapUtils, CommandUtils, ConfigLoader, CryptoUtils, RemoteNodeService, VotingService } from '../service';
 
 const logger: Logger = LoggerFactory.getLogger(LogType.System);
@@ -57,16 +55,24 @@ When a new voting file is created, Bootstrap will advise running the \`link\` co
         const target = flags.target;
         const configLoader = new ConfigLoader();
         const addressesLocation = configLoader.getGeneratedAddressLocation(target);
-        const existingPreset = configLoader.loadExistingPresetData(target, password);
-        const preset = existingPreset.preset;
-        if (!preset) {
-            throw new Error(`Network preset could not be resolved!`);
-        }
-        // Adds new shared/network properties to the existing preset. This is for upgrades..
-        const sharedPreset = BootstrapUtils.loadYaml(join(BootstrapUtils.ROOT_FOLDER, 'presets', 'shared.yml'), false);
-        const networkPreset = BootstrapUtils.loadYaml(join(BootstrapUtils.ROOT_FOLDER, 'presets', preset, 'network.yml'), false);
-        const presetData = configLoader.mergePresets(sharedPreset, networkPreset, existingPreset) as ConfigPreset;
 
+        const loadPresetData = () => {
+            try {
+                const oldPresetData = configLoader.loadExistingPresetData(target, password);
+                return configLoader.createPresetData({
+                    password: password,
+                    oldPresetData,
+                });
+            } catch (e) {
+                throw new Error(
+                    `Node's preset cannot be loaded. Have you provided the right --target? If you have, please rerun the 'config' command with --upgrade. Error: ${
+                        e.message || 'unknown'
+                    }`,
+                );
+            }
+        };
+
+        const presetData = loadPresetData();
         const addresses = configLoader.loadExistingAddresses(target, password);
         const privateKeySecurityMode = CryptoUtils.getPrivateKeySecurityMode(presetData.privateKeySecurityMode);